HPCC-24425 DESDL: preserve scriptContext across entire transaction
We will be using the scriptContext for multiple new features and need it to exist for the entire transaction.

Including that the scriptContext will exist for transaction logging at the end of the transaction.

Provide a low level mechanism where the script can skip a particular log thread queue entry.
Example:
 <xsdl:set-log-option name="thread-queue-myserv-agent1" select="'skip'"/>

Signed-off-by: Anthony Fishbeck <anthony.fishbeck@lexisnexisrisk.com>